перенос моделей и моделей-представлния в проект Предметной области
