+++ Common/WebStore.Domain/Models/CartItem.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace WebStore.Models
+namespace WebStore.Domain.Models
 {
     public class CartItem
     {
