+++ Common/WebStore.Domain/ViewModels/CartOrderViewModel.cs
@@ -2,9 +2,9 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using WebStore.ViewModels.Orders;
+using WebStore.Domain.ViewModels.Orders;
 
-namespace WebStore.ViewModels
+namespace WebStore.Domain.ViewModels
 {
     public class CartOrderViewModel
     {
