+++ Common/WebStore.Domain/ViewModels/Identity/RegisterUserViewModel.cs
