+++ Common/WebStore.Domain/ViewModels/Orders/OrderViewModel.cs
@@ -4,7 +4,7 @@ using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace WebStore.ViewModels.Orders
+namespace WebStore.Domain.ViewModels.Orders
 {
     public class OrderViewModel
     {
