+++ Common/WebStore.Domain/ViewModels/Orders/UserOrderViewModel.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace WebStore.ViewModels.Orders
+namespace WebStore.Domain.ViewModels.Orders
 {
     public class UserOrderViewModel
     {
