+++ Common/WebStore.Domain/ViewModels/ProductViewModel.cs
@@ -4,12 +4,15 @@ using System.Linq;
 using System.Threading.Tasks;
 using WebStore.Domain.Entities.Base.Interfaces;
 
-namespace WebStore.ViewModels
+namespace WebStore.Domain.ViewModels
 {
-    public class BrandViewModel : INamedEntity, IOrderedEntity
+    public class ProductViewModel : INamedEntity, IOrderedEntity
     {
         public string Name { get; set; }
         public int Id { get; set; }
         public int Order { get; set; }
+        public string ImageUrl { get; set; }
+        public decimal Price { get; set; }
+        public string? Brand { get; set; }
     }
 }
